Move start_query_execution back
