+++ awswrangler/athena/__init__.py
@@ -4,7 +4,6 @@ from awswrangler.athena._read import (  # noqa
     get_query_results,
     read_sql_query,
     read_sql_table,
-    start_query_execution,
     unload,
 )
 from awswrangler.athena._utils import (  # noqa
@@ -20,6 +19,7 @@ from awswrangler.athena._utils import (  # noqa
     list_query_executions,
     repair_table,
     show_create_table,
+    start_query_execution,
     stop_query_execution,
     wait_query,
 )
